added the deleting process in neo4j with consistency with other db , resolving naming and type issue, like enginedisplacement=double and added annotation @field  to carbrand name and model

+++ src/main/java/it/unipi/CarRev/service/Impl/DeleteACarServiceImpl.java
@@ -3,14 +3,18 @@ package it.unipi.CarRev.service.Impl;
 
 import it.unipi.CarRev.config.RedisConfig;
 import it.unipi.CarRev.dao.mongo.CarDAO;
+import it.unipi.CarRev.model.Car;
 import org.springframework.stereotype.Service;
+import it.unipi.CarRev.service.Neo4jCarDeleteService;
 import redis.clients.jedis.Jedis;
 
 @Service
 public class DeleteACarServiceImpl {
     private final CarDAO carDAO;
-    public DeleteACarServiceImpl(CarDAO carDAO){
+    private final Neo4jCarDeleteService neo4jCarDeleteService ;
+    public DeleteACarServiceImpl(CarDAO carDAO,Neo4jCarDeleteService neo4jCarDeleteService){
         this.carDAO=carDAO;
+        this.neo4jCarDeleteService = neo4jCarDeleteService;
     }
 
     public int deleteCar(String id){
@@ -19,13 +23,26 @@ public class DeleteACarServiceImpl {
                 System.out.println("car not found");
                 return -1;
             }
+            Car oldCar = carDAO.findById(id).orElse(null);
+            if (oldCar == null) {
+                System.out.println("car not found (race condition)");
+                return -1;
+            }
+
+            int deleted = neo4jCarDeleteService.deleteCarProjection(oldCar);
+            if (deleted != 1) {
+                System.out.println("neo4j mismatch on delete: deleted=" + deleted);
+                return -2; //
+            }
             carDAO.deleteById(id);
             int result=deleteInRedis(id);
             return result;
 
         }
         catch(Exception e){
+
             System.out.println("an error has occurred during the delete of a car");
+            e.printStackTrace();
             return -2;
         }
     }
